scripts/requirements.txt
python-pptx
beautifulsoup
pillow